Fix: preserve input dtype in l2norm helper

+++ flash_qla/utils/math.py
@@ -7,7 +7,7 @@ import torch
 @torch.compile
 def l2norm_compiled(x: torch.FloatTensor, dim: int = -1, eps: float = 1e-6):
     inv_norm = torch.rsqrt((x * x).sum(dim=dim, keepdim=True) + eps)
-    return x * inv_norm
+    return (x * inv_norm).to(x.dtype)
 
 
 def l2norm(x: torch.FloatTensor, dim: int = -1, eps: float = 1e-6):
